Added in support keychains in general and OS X specifically.
The framework for keychains is now added in.  The config file now has a
new parameter which is checked for properly in the config file reading
part of putmail.  The accepted values are coded into the keychain()
function.  The only working value now is 'osx'.  This option calls the
osxkeychain() function, which in turn calls upon the OS X 'security'
program, using the defined server as it's argument.

This framework should be easily expanded by including other keychain
functions.  Any keychain which is accessible from the command line
should be easily parsed for it's output.